% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltAnomalies.R
\name{as.data.table.MiltAnomalies}
\alias{as.data.table.MiltAnomalies}
\title{Convert MiltAnomalies to a data.table}
\usage{
\method{as.data.table}{MiltAnomalies}(x, ...)
}
\arguments{
\item{x}{A \code{MiltAnomalies} object.}

\item{...}{Ignored.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} with the same columns as
\code{\link[=as_tibble.MiltAnomalies]{as_tibble.MiltAnomalies()}}.
}
\description{
Convert MiltAnomalies to a data.table
}
