% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltForecast.R
\name{as.data.table.MiltForecast}
\alias{as.data.table.MiltForecast}
\title{Convert a MiltForecast to a data.table}
\usage{
\method{as.data.table}{MiltForecast}(x, ...)
}
\arguments{
\item{x}{A \code{MiltForecast} object.}

\item{...}{Ignored.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}} with the same columns as
\code{\link[=as_tibble.MiltForecast]{as_tibble.MiltForecast()}}.
}
\description{
Convert a MiltForecast to a data.table
}
