% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltSeries.R
\name{as.data.table.MiltSeries}
\alias{as.data.table.MiltSeries}
\title{Convert a MiltSeries to a data.table}
\usage{
\method{as.data.table}{MiltSeries}(x, ...)
}
\arguments{
\item{x}{A \code{MiltSeries} object.}

\item{...}{Ignored.}
}
\value{
A \code{\link[data.table:data.table]{data.table::data.table()}}.
}
\description{
Convert a MiltSeries to a data.table
}
