% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltAnomalies.R
\name{as_tibble.MiltAnomalies}
\alias{as_tibble.MiltAnomalies}
\title{Convert MiltAnomalies to tibble}
\usage{
\method{as_tibble}{MiltAnomalies}(x, ...)
}
\arguments{
\item{x}{A \code{MiltAnomalies} object.}

\item{...}{Ignored.}
}
\value{
A \code{\link[tibble:tibble]{tibble::tibble()}} with columns \code{time}, \code{value}, \code{.is_anomaly},
\code{.anomaly_score}, and compatibility aliases \code{is_anomaly},
\code{anomaly_score}.
}
\description{
Convert MiltAnomalies to tibble
}
