% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{get_milt_model_class}
\alias{get_milt_model_class}
\title{Retrieve a registered model's R6 class generator}
\usage{
get_milt_model_class(name)
}
\arguments{
\item{name}{Character scalar. Model identifier.}
}
\value{
The R6 class generator for the requested model.
}
\description{
Retrieve a registered model's R6 class generator
}
\keyword{internal}
