% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{is_registered_model}
\alias{is_registered_model}
\title{Check whether a model name is registered}
\usage{
is_registered_model(name)
}
\arguments{
\item{name}{Character scalar.}
}
\value{
\code{TRUE} / \code{FALSE}.
}
\description{
Check whether a model name is registered
}
\keyword{internal}
