% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-registry.R
\name{list_milt_models}
\alias{list_milt_models}
\title{List all registered milt models}
\usage{
list_milt_models()
}
\value{
A tibble with columns \code{name}, \code{description}, \code{multivariate},
\code{probabilistic}, \code{covariates}, \code{multi_series}.
}
\description{
List all registered milt models
}
\examples{
list_milt_models()
}
