% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomaly-model.R
\name{milt_detect_anomalies}
\alias{milt_detect_anomalies}
\title{Detect anomalies from a fitted MiltAnomalyModel}
\usage{
milt_detect_anomalies(model)
}
\arguments{
\item{model}{A fitted \code{MiltAnomalyModel}, created by \code{\link[=milt_anomaly_model]{milt_anomaly_model()}}
and fit with \code{\link[=milt_fit]{milt_fit()}}.}
}
\value{
A \code{MiltAnomalies} object.
}
\description{
Detect anomalies from a fitted MiltAnomalyModel
}
\seealso{
\code{\link[=milt_anomaly_model]{milt_anomaly_model()}}

Other anomaly:
\code{\link[=milt_anomaly_model]{milt_anomaly_model()}},
\code{\link[=milt_anomaly_score]{milt_anomaly_score()}},
\code{\link[=milt_causal_impact]{milt_causal_impact()}},
\code{\link[=milt_changepoints]{milt_changepoints()}},
\code{\link[=milt_detect]{milt_detect()}},
\code{\link[=milt_detector]{milt_detector()}}
}
\concept{anomaly}
