% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{milt_rmse}
\alias{milt_rmse}
\title{Root Mean Squared Error}
\usage{
milt_rmse(actual, predicted)
}
\arguments{
\item{actual}{Numeric vector of observed values.}

\item{predicted}{Numeric vector of predicted values.}
}
\value{
A single numeric value.
}
\description{
Root Mean Squared Error
}
\seealso{
Other metrics:
\code{\link[=milt_accuracy]{milt_accuracy()}},
\code{\link[=milt_coefficient_of_variation]{milt_coefficient_of_variation()}},
\code{\link[=milt_coverage]{milt_coverage()}},
\code{\link[=milt_crps]{milt_crps()}},
\code{\link[=milt_mae]{milt_mae()}},
\code{\link[=milt_mape]{milt_mape()}},
\code{\link[=milt_marre]{milt_marre()}},
\code{\link[=milt_mase]{milt_mase()}},
\code{\link[=milt_mrae]{milt_mrae()}},
\code{\link[=milt_mse]{milt_mse()}},
\code{\link[=milt_ope]{milt_ope()}},
\code{\link[=milt_pinball]{milt_pinball()}},
\code{\link[=milt_r_squared]{milt_r_squared()}},
\code{\link[=milt_rmsle]{milt_rmsle()}},
\code{\link[=milt_rmsse]{milt_rmsse()}},
\code{\link[=milt_smape]{milt_smape()}},
\code{\link[=milt_winkler]{milt_winkler()}},
\code{\link[=milt_wmape]{milt_wmape()}}
}
\concept{metrics}
