% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-transform.R
\name{milt_step_unboxcox}
\alias{milt_step_unboxcox}
\title{Invert a Box-Cox transform on a time series}
\usage{
milt_step_unboxcox(step, series)
}
\arguments{
\item{step}{A \code{MiltBoxCoxStep} object returned by \code{\link[=milt_step_boxcox]{milt_step_boxcox()}}.}

\item{series}{A \code{MiltSeries} to invert.}
}
\value{
The original-scale \code{MiltSeries}.
}
\description{
Convenience wrapper around \code{MiltBoxCoxStep$inverse_transform()}.
}
\seealso{
\code{\link[=milt_step_boxcox]{milt_step_boxcox()}}

Other features:
\code{\link[=milt_step_boxcox]{milt_step_boxcox()}},
\code{\link[=milt_step_calendar]{milt_step_calendar()}},
\code{\link[=milt_step_diff]{milt_step_diff()}},
\code{\link[=milt_step_fourier]{milt_step_fourier()}},
\code{\link[=milt_step_lag]{milt_step_lag()}},
\code{\link[=milt_step_map]{milt_step_map()}},
\code{\link[=milt_step_rolling]{milt_step_rolling()}},
\code{\link[=milt_step_scale]{milt_step_scale()}},
\code{\link[=milt_step_undiff]{milt_step_undiff()}},
\code{\link[=milt_step_unscale]{milt_step_unscale()}}
}
\concept{features}
