% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-scale.R
\name{milt_step_unscale}
\alias{milt_step_unscale}
\title{Invert a scaling step on a time series}
\usage{
milt_step_unscale(step, series)
}
\arguments{
\item{step}{A \code{MiltScaleStep} object returned by \code{\link[=milt_step_scale]{milt_step_scale()}}.}

\item{series}{A \code{MiltSeries} to unscale.}
}
\value{
The unscaled \code{MiltSeries}.
}
\description{
Convenience wrapper around \code{MiltScaleStep$inverse_transform()}.
}
\seealso{
\code{\link[=milt_step_scale]{milt_step_scale()}}

Other features:
\code{\link[=milt_step_boxcox]{milt_step_boxcox()}},
\code{\link[=milt_step_calendar]{milt_step_calendar()}},
\code{\link[=milt_step_diff]{milt_step_diff()}},
\code{\link[=milt_step_fourier]{milt_step_fourier()}},
\code{\link[=milt_step_lag]{milt_step_lag()}},
\code{\link[=milt_step_map]{milt_step_map()}},
\code{\link[=milt_step_rolling]{milt_step_rolling()}},
\code{\link[=milt_step_scale]{milt_step_scale()}},
\code{\link[=milt_step_unboxcox]{milt_step_unboxcox()}},
\code{\link[=milt_step_undiff]{milt_step_undiff()}}
}
\concept{features}
