% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-manipulate.R
\name{milt_tail}
\alias{milt_tail}
\title{Return the last n observations of a MiltSeries}
\usage{
milt_tail(series, n = 6L)
}
\arguments{
\item{series}{A \code{MiltSeries} object.}

\item{n}{Number of observations to return. Default 6.}
}
\value{
A \code{MiltSeries}.
}
\description{
Return the last n observations of a MiltSeries
}
\seealso{
Other series:
\code{\link[=milt_add_covariates]{milt_add_covariates()}},
\code{\link[=milt_add_datetime_component]{milt_add_datetime_component()}},
\code{\link[=milt_add_holidays]{milt_add_holidays()}},
\code{\link[=milt_check_seasonality]{milt_check_seasonality()}},
\code{\link[=milt_concat]{milt_concat()}},
\code{\link[=milt_diagnose]{milt_diagnose()}},
\code{\link[=milt_fill_gaps]{milt_fill_gaps()}},
\code{\link[=milt_filter]{milt_filter()}},
\code{\link[=milt_get_covariates]{milt_get_covariates()}},
\code{\link[=milt_head]{milt_head()}},
\code{\link[=milt_plot_acf]{milt_plot_acf()}},
\code{\link[=milt_plot_decomp]{milt_plot_decomp()}},
\code{\link[=milt_read_parquet]{milt_read_parquet()}},
\code{\link[=milt_resample]{milt_resample()}},
\code{\link[=milt_series]{milt_series()}},
\code{\link[=milt_split]{milt_split()}},
\code{\link[=milt_split_at]{milt_split_at()}},
\code{\link[=milt_stack]{milt_stack()}},
\code{\link[=milt_window]{milt_window()}},
\code{\link[=milt_write_parquet]{milt_write_parquet()}},
\code{\link[=plot.MiltSeries]{plot.MiltSeries()}}
}
\concept{series}
