% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversions.R
\name{milt_to_tibble}
\alias{milt_to_tibble}
\title{Convert a MiltSeries to a tibble}
\usage{
milt_to_tibble(series)
}
\arguments{
\item{series}{A \code{MiltSeries} object.}
}
\value{
A \code{tibble}.
}
\description{
Convert a MiltSeries to a tibble
}
