% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversions.R
\name{milt_to_ts}
\alias{milt_to_ts}
\title{Convert a MiltSeries to a base R ts object}
\usage{
milt_to_ts(series)
}
\arguments{
\item{series}{A \code{MiltSeries} object. Must be univariate.}
}
\value{
A \code{ts} object.
}
\description{
Convert a MiltSeries to a base R ts object
}
