% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-conversions.R
\name{milt_to_tsibble}
\alias{milt_to_tsibble}
\title{Convert a MiltSeries to a tsibble}
\usage{
milt_to_tsibble(series)
}
\arguments{
\item{series}{A \code{MiltSeries} object.}
}
\value{
A \code{tsibble}.
}
\description{
Convert a MiltSeries to a tsibble
}
