% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-reticulate.R
\name{MiltDartsRNN}
\alias{MiltDartsRNN}
\title{Darts RNN model (LSTM / GRU)}
\description{
Wraps Darts' \code{RNNModel} via \code{reticulate}.  Requires Python and the \code{darts}
package; call \code{\link[=milt_setup_darts]{milt_setup_darts()}} first.
}
\section{Super classes}{
\code{milt::MiltModel} -> \code{milt::MiltDartsModel} -> \code{MiltDartsRNN}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MiltDartsRNN-new}{\code{MiltDartsRNN$new()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltModel" data-id="get_params"><a href='../../milt/html/MiltModel.html#method-MiltModel-get_params'><code>milt::MiltModel$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltModel" data-id="is_fitted"><a href='../../milt/html/MiltModel.html#method-MiltModel-is_fitted'><code>milt::MiltModel$is_fitted()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltModel" data-id="summary"><a href='../../milt/html/MiltModel.html#method-MiltModel-summary'><code>milt::MiltModel$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltDartsModel" data-id="fit"><a href='../../milt/html/MiltDartsModel.html#method-MiltDartsModel-fit'><code>milt::MiltDartsModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltDartsModel" data-id="forecast"><a href='../../milt/html/MiltDartsModel.html#method-MiltDartsModel-forecast'><code>milt::MiltDartsModel$forecast()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltDartsModel" data-id="predict"><a href='../../milt/html/MiltDartsModel.html#method-MiltDartsModel-predict'><code>milt::MiltDartsModel$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltDartsModel" data-id="residuals"><a href='../../milt/html/MiltDartsModel.html#method-MiltDartsModel-residuals'><code>milt::MiltDartsModel$residuals()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltDartsRNN-new"></a>}}
\if{latex}{\out{\hypertarget{method-MiltDartsRNN-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltDartsRNN$new(
  model = "LSTM",
  hidden_dim = 64L,
  n_rnn_layers = 2L,
  input_chunk_length = 24L,
  training_length = 36L,
  n_epochs = 100L,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{Character. RNN cell type: \code{"LSTM"} (default) or \code{"GRU"}.}

\item{\code{hidden_dim}}{Integer. Hidden size. Default \code{64L}.}

\item{\code{n_rnn_layers}}{Integer. Number of recurrent layers. Default \code{2L}.}

\item{\code{input_chunk_length}}{Integer. Lookback window. Default \code{24L}.}

\item{\code{training_length}}{Integer. Length of sequences used in training.
Must be \verb{> input_chunk_length}. Default \code{36L}.}

\item{\code{n_epochs}}{Integer. Training epochs. Default \code{100L}.}

\item{\code{...}}{Further arguments passed to \code{darts.models.RNNModel}.}
}
\if{html}{\out{</div>}}
}
}
}
