% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dl-reticulate.R
\name{MiltDartsTransformer}
\alias{MiltDartsTransformer}
\title{Darts Transformer model}
\description{
Wraps Darts' \code{TransformerModel} via \code{reticulate}.  Requires Python and the
\code{darts} package; call \code{\link[=milt_setup_darts]{milt_setup_darts()}} first.
}
\section{Super classes}{
\code{milt::MiltModel} -> \code{milt::MiltDartsModel} -> \code{MiltDartsTransformer}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MiltDartsTransformer-new}{\code{MiltDartsTransformer$new()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltModel" data-id="get_params"><a href='../../milt/html/MiltModel.html#method-MiltModel-get_params'><code>milt::MiltModel$get_params()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltModel" data-id="is_fitted"><a href='../../milt/html/MiltModel.html#method-MiltModel-is_fitted'><code>milt::MiltModel$is_fitted()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltModel" data-id="summary"><a href='../../milt/html/MiltModel.html#method-MiltModel-summary'><code>milt::MiltModel$summary()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltDartsModel" data-id="fit"><a href='../../milt/html/MiltDartsModel.html#method-MiltDartsModel-fit'><code>milt::MiltDartsModel$fit()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltDartsModel" data-id="forecast"><a href='../../milt/html/MiltDartsModel.html#method-MiltDartsModel-forecast'><code>milt::MiltDartsModel$forecast()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltDartsModel" data-id="predict"><a href='../../milt/html/MiltDartsModel.html#method-MiltDartsModel-predict'><code>milt::MiltDartsModel$predict()</code></a></span></li>
<li><span class="pkg-link" data-pkg="milt" data-topic="MiltDartsModel" data-id="residuals"><a href='../../milt/html/MiltDartsModel.html#method-MiltDartsModel-residuals'><code>milt::MiltDartsModel$residuals()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltDartsTransformer-new"></a>}}
\if{latex}{\out{\hypertarget{method-MiltDartsTransformer-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltDartsTransformer$new(
  input_chunk_length = 24L,
  output_chunk_length = 12L,
  d_model = 64L,
  nhead = 4L,
  num_encoder_layers = 2L,
  num_decoder_layers = 2L,
  n_epochs = 100L,
  ...
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input_chunk_length}}{Integer. Encoder lookback. Default \code{24L}.}

\item{\code{output_chunk_length}}{Integer. Decoder output length. Default \code{12L}.}

\item{\code{d_model}}{Integer. Model dimension. Default \code{64L}.}

\item{\code{nhead}}{Integer. Attention heads. Default \code{4L}.}

\item{\code{num_encoder_layers}}{Integer. Encoder depth. Default \code{2L}.}

\item{\code{num_decoder_layers}}{Integer. Decoder depth. Default \code{2L}.}

\item{\code{n_epochs}}{Integer. Training epochs. Default \code{100L}.}

\item{\code{...}}{Further arguments passed to \code{darts.models.TransformerModel}.}
}
\if{html}{\out{</div>}}
}
}
}
