% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect-interface.R
\name{MiltDetectorBase}
\alias{MiltDetectorBase}
\title{Run detection on a MiltSeries. Override in each backend.}
\description{
Run detection on a MiltSeries. Override in each backend.

Run detection on a MiltSeries. Override in each backend.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MiltDetector-new}{\code{MiltDetectorBase$new()}}
\item \href{#method-MiltDetector-detect}{\code{MiltDetectorBase$detect()}}
\item \href{#method-MiltDetector-name}{\code{MiltDetectorBase$name()}}
\item \href{#method-MiltDetector-get_params}{\code{MiltDetectorBase$get_params()}}
\item \href{#method-MiltDetector-clone}{\code{MiltDetectorBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltDetector-new"></a>}}
\if{latex}{\out{\hypertarget{method-MiltDetector-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltDetectorBase$new(name = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltDetector-detect"></a>}}
\if{latex}{\out{\hypertarget{method-MiltDetector-detect}{}}}
\subsection{Method \code{detect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltDetectorBase$detect(series, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltDetector-name"></a>}}
\if{latex}{\out{\hypertarget{method-MiltDetector-name}{}}}
\subsection{Method \code{name()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltDetectorBase$name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The detector name string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltDetector-get_params"></a>}}
\if{latex}{\out{\hypertarget{method-MiltDetector-get_params}{}}}
\subsection{Method \code{get_params()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltDetectorBase$get_params()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The hyperparameter list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltDetector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MiltDetector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltDetectorBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
