% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features-scale.R
\name{MiltScaleStepR6}
\alias{MiltScaleStepR6}
\title{Apply the learned scaling to a new series.
Reverse the scaling on a series.
Reverse-scale a numeric vector using the FIRST value column's parameters.}
\description{
Apply the learned scaling to a new series.
Reverse the scaling on a series.
Reverse-scale a numeric vector using the FIRST value column's parameters.

Apply the learned scaling to a new series.
Reverse the scaling on a series.
Reverse-scale a numeric vector using the FIRST value column's parameters.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MiltScaleStep-new}{\code{MiltScaleStepR6$new()}}
\item \href{#method-MiltScaleStep-method}{\code{MiltScaleStepR6$method()}}
\item \href{#method-MiltScaleStep-center}{\code{MiltScaleStepR6$center()}}
\item \href{#method-MiltScaleStep-scale}{\code{MiltScaleStepR6$scale()}}
\item \href{#method-MiltScaleStep-transform}{\code{MiltScaleStepR6$transform()}}
\item \href{#method-MiltScaleStep-inverse_transform}{\code{MiltScaleStepR6$inverse_transform()}}
\item \href{#method-MiltScaleStep-inverse_transform_vector}{\code{MiltScaleStepR6$inverse_transform_vector()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltScaleStep-new"></a>}}
\if{latex}{\out{\hypertarget{method-MiltScaleStep-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltScaleStepR6$new(method, center, scale)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltScaleStep-method"></a>}}
\if{latex}{\out{\hypertarget{method-MiltScaleStep-method}{}}}
\subsection{Method \code{method()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltScaleStepR6$method()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltScaleStep-center"></a>}}
\if{latex}{\out{\hypertarget{method-MiltScaleStep-center}{}}}
\subsection{Method \code{center()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltScaleStepR6$center()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltScaleStep-scale"></a>}}
\if{latex}{\out{\hypertarget{method-MiltScaleStep-scale}{}}}
\subsection{Method \code{scale()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltScaleStepR6$scale()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltScaleStep-transform"></a>}}
\if{latex}{\out{\hypertarget{method-MiltScaleStep-transform}{}}}
\subsection{Method \code{transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltScaleStepR6$transform(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltScaleStep-inverse_transform"></a>}}
\if{latex}{\out{\hypertarget{method-MiltScaleStep-inverse_transform}{}}}
\subsection{Method \code{inverse_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltScaleStepR6$inverse_transform(series)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MiltScaleStep-inverse_transform_vector"></a>}}
\if{latex}{\out{\hypertarget{method-MiltScaleStep-inverse_transform_vector}{}}}
\subsection{Method \code{inverse_transform_vector()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MiltScaleStepR6$inverse_transform_vector(x, col = NULL)}\if{html}{\out{</div>}}
}

}
}
