% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltAnomalies.R
\name{plot.MiltAnomalies}
\alias{plot.MiltAnomalies}
\title{Plot a MiltAnomalies object}
\usage{
\method{plot}{MiltAnomalies}(x, ...)
}
\arguments{
\item{x}{A \code{MiltAnomalies} object.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot2} plot object.
}
\description{
Draws the time series with anomalous points highlighted in red.
}
