% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltAnomalies.R
\name{print.MiltAnomalies}
\alias{print.MiltAnomalies}
\title{Print a MiltAnomalies object}
\usage{
\method{print}{MiltAnomalies}(x, ...)
}
\arguments{
\item{x}{A \code{MiltAnomalies} object.}

\item{...}{Ignored.}
}
\description{
Print a MiltAnomalies object
}
