% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{print.MiltBacktest}
\alias{print.MiltBacktest}
\title{Print a MiltBacktest}
\usage{
\method{print}{MiltBacktest}(x, ...)
}
\arguments{
\item{x}{A \code{MiltBacktest} object.}

\item{...}{Ignored.}
}
\description{
Print a MiltBacktest
}
