% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{print.MiltComparison}
\alias{print.MiltComparison}
\title{Print a MiltComparison}
\usage{
\method{print}{MiltComparison}(x, ...)
}
\arguments{
\item{x}{A \code{MiltComparison} object.}

\item{...}{Ignored.}
}
\description{
Print a MiltComparison
}
