% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltForecast.R
\name{print.MiltForecast}
\alias{print.MiltForecast}
\title{Print a MiltForecast}
\usage{
\method{print}{MiltForecast}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{MiltForecast} object.}

\item{n}{Number of rows to preview. Default 6.}

\item{...}{Ignored.}
}
\description{
Print a MiltForecast
}
