% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-interface.R
\name{print.MiltModel}
\alias{print.MiltModel}
\title{Print a MiltModel}
\usage{
\method{print}{MiltModel}(x, ...)
}
\arguments{
\item{x}{A \code{MiltModel} object.}

\item{...}{Ignored.}
}
\description{
Print a MiltModel
}
