% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltSeries.R
\name{print.MiltSeries}
\alias{print.MiltSeries}
\title{Print a MiltSeries}
\usage{
\method{print}{MiltSeries}(x, n = 6L, ...)
}
\arguments{
\item{x}{A \code{MiltSeries} object.}

\item{n}{Number of rows to preview. Default 6.}

\item{...}{Ignored.}
}
\description{
Print a MiltSeries
}
