% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltSeries.R
\name{[.MiltSeries}
\alias{[.MiltSeries}
\title{Subset a MiltSeries by row index}
\usage{
\method{[}{MiltSeries}(x, i, ...)
}
\arguments{
\item{x}{A \code{MiltSeries} object.}

\item{i}{Integer index, logical vector, or a length-2 Date/POSIXct vector
specifying \code{c(start, end)}.}

\item{...}{Ignored.}
}
\description{
Subset a MiltSeries by row index
}
