% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltAnomalies.R
\name{summary.MiltAnomalies}
\alias{summary.MiltAnomalies}
\title{Summarise a MiltAnomalies object}
\usage{
\method{summary}{MiltAnomalies}(object, ...)
}
\arguments{
\item{object}{A \code{MiltAnomalies} object.}

\item{...}{Ignored.}
}
\description{
Summarise a MiltAnomalies object
}
