% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{summary.MiltBacktest}
\alias{summary.MiltBacktest}
\title{Summary of a MiltBacktest}
\usage{
\method{summary}{MiltBacktest}(object, ...)
}
\arguments{
\item{object}{A \code{MiltBacktest} object.}

\item{...}{Ignored.}
}
\description{
Summary of a MiltBacktest
}
