% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{summary.MiltComparison}
\alias{summary.MiltComparison}
\title{Summary of a MiltComparison}
\usage{
\method{summary}{MiltComparison}(object, ...)
}
\arguments{
\item{object}{A \code{MiltComparison} object.}

\item{...}{Ignored.}
}
\description{
Summary of a MiltComparison
}
