% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltForecast.R
\name{summary.MiltForecast}
\alias{summary.MiltForecast}
\title{Summarise a MiltForecast}
\usage{
\method{summary}{MiltForecast}(object, ...)
}
\arguments{
\item{object}{A \code{MiltForecast} object.}

\item{...}{Ignored.}
}
\description{
Summarise a MiltForecast
}
