% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiltSeries.R
\name{summary.MiltSeries}
\alias{summary.MiltSeries}
\title{Summarise a MiltSeries}
\usage{
\method{summary}{MiltSeries}(object, ...)
}
\arguments{
\item{object}{A \code{MiltSeries} object.}

\item{...}{Ignored.}
}
\description{
Returns descriptive statistics for each value column.
}
