% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{milt_classify_fit}
\alias{milt_classify_fit}
\title{Fit a time series classifier}
\usage{
milt_classify_fit(classifier, series_list, labels)
}
\arguments{
\item{classifier}{A \code{MiltClassifier} from \code{\link[=milt_classifier]{milt_classifier()}}.}

\item{series_list}{A list of \code{MiltSeries} objects (training set).}

\item{labels}{Character or factor vector of class labels, one per series.}
}
\value{
The fitted \code{MiltClassifier} (invisibly, mutated in place).
}
\description{
Trains the classifier on a labelled set of time series.
}
\seealso{
\code{\link[=milt_classifier]{milt_classifier()}}, \code{\link[=milt_classify_predict]{milt_classify_predict()}}

Other classify: 
\code{\link{milt_classifier}()},
\code{\link{milt_classify_predict}()}
}
\concept{classify}
