% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-covariates.R
\name{milt_get_covariates}
\alias{milt_get_covariates}
\title{Retrieve covariates attached to a MiltSeries}
\usage{
milt_get_covariates(series, type = "past")
}
\arguments{
\item{series}{A \code{MiltSeries} object.}

\item{type}{One of \code{"past"}, \code{"future"}, or \code{"static"}. Default \code{"past"}.}
}
\value{
A tibble of covariates, or \code{NULL} if none have been attached.
}
\description{
Retrieve covariates attached to a MiltSeries
}
\seealso{
Other series: 
\code{\link{milt_add_covariates}()},
\code{\link{milt_concat}()},
\code{\link{milt_diagnose}()},
\code{\link{milt_fill_gaps}()},
\code{\link{milt_head}()},
\code{\link{milt_plot_acf}()},
\code{\link{milt_plot_decomp}()},
\code{\link{milt_resample}()},
\code{\link{milt_series}()},
\code{\link{milt_split}()},
\code{\link{milt_split_at}()},
\code{\link{milt_tail}()},
\code{\link{milt_window}()},
\code{\link{plot.MiltSeries}()}
}
\concept{series}
