% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{milt_mae}
\alias{milt_mae}
\title{Mean Absolute Error}
\usage{
milt_mae(actual, predicted)
}
\arguments{
\item{actual}{Numeric vector of observed values.}

\item{predicted}{Numeric vector of predicted values.}
}
\value{
A single numeric value.
}
\description{
Mean Absolute Error
}
\seealso{
Other metrics: 
\code{\link{milt_accuracy}()},
\code{\link{milt_coverage}()},
\code{\link{milt_crps}()},
\code{\link{milt_mape}()},
\code{\link{milt_mase}()},
\code{\link{milt_mrae}()},
\code{\link{milt_mse}()},
\code{\link{milt_pinball}()},
\code{\link{milt_r_squared}()},
\code{\link{milt_rmse}()},
\code{\link{milt_rmsse}()},
\code{\link{milt_smape}()},
\code{\link{milt_winkler}()}
}
\concept{metrics}
