% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{milt_pipeline}
\alias{milt_pipeline}
\title{Create a new milt pipeline}
\usage{
milt_pipeline()
}
\value{
An empty \code{MiltPipeline} object.
}
\description{
Returns an empty \code{MiltPipeline} object. Chain steps onto it with
\verb{milt_pipe_step_*()} functions and attach a model with \code{\link[=milt_pipe_model]{milt_pipe_model()}}.
Fit the whole pipeline with \code{\link[=milt_pipeline_fit]{milt_pipeline_fit()}}.
}
\examples{
pipe <- milt_pipeline()
pipe
}
\seealso{
\code{\link[=milt_pipeline_fit]{milt_pipeline_fit()}}, \code{\link[=milt_pipeline_forecast]{milt_pipeline_forecast()}},
\code{\link[=milt_pipe_step_lag]{milt_pipe_step_lag()}}, \code{\link[=milt_pipe_step_scale]{milt_pipe_step_scale()}},
\code{\link[=milt_pipe_step_rolling]{milt_pipe_step_rolling()}}, \code{\link[=milt_pipe_model]{milt_pipe_model()}}

Other pipeline: 
\code{\link{milt_pipe_model}()},
\code{\link{milt_pipe_step_calendar}()},
\code{\link{milt_pipe_step_fourier}()},
\code{\link{milt_pipe_step_lag}()},
\code{\link{milt_pipe_step_rolling}()},
\code{\link{milt_pipe_step_scale}()},
\code{\link{milt_pipeline_fit}()},
\code{\link{milt_pipeline_forecast}()},
\code{\link{milt_pipeline_transform}()}
}
\concept{pipeline}
