% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/series-manipulate.R
\name{milt_tail}
\alias{milt_tail}
\title{Return the last n observations of a MiltSeries}
\usage{
milt_tail(series, n = 6L)
}
\arguments{
\item{series}{A \code{MiltSeries} object.}

\item{n}{Number of observations to return. Default 6.}
}
\value{
A \code{MiltSeries}.
}
\description{
Return the last n observations of a MiltSeries
}
\seealso{
Other series: 
\code{\link{milt_add_covariates}()},
\code{\link{milt_concat}()},
\code{\link{milt_diagnose}()},
\code{\link{milt_fill_gaps}()},
\code{\link{milt_get_covariates}()},
\code{\link{milt_head}()},
\code{\link{milt_plot_acf}()},
\code{\link{milt_plot_decomp}()},
\code{\link{milt_resample}()},
\code{\link{milt_series}()},
\code{\link{milt_split}()},
\code{\link{milt_split_at}()},
\code{\link{milt_window}()},
\code{\link{plot.MiltSeries}()}
}
\concept{series}
